separated relay namespace and hybrid_connection ID parse and validate
