+++ azurerm/internal/services/relay/relay_namespace_test.go
@@ -4,57 +4,6 @@ import (
 	"testing"
 )
 
-func TestParseHybridConnectionID(t *testing.T) {
-	testData := []struct {
-		Name     string
-		Input    string
-		Expected *HybridConnectionResourceID
-	}{
-		{
-			Name:     "Empty",
-			Input:    "",
-			Expected: nil,
-		},
-		{
-			Name:     "No Resource Groups Segment",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000",
-			Expected: nil,
-		},
-		{
-			Name:     "No Resource Groups Value",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/",
-			Expected: nil,
-		},
-		{
-			Name:     "Resource Group ID",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/",
-			Expected: nil,
-		},
-		// Todo: complete coverage
-	}
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Name)
-
-		actual, err := ParseHybridConnectionID(v.Input)
-		if err != nil {
-			if v.Expected == nil {
-				continue
-			}
-
-			t.Fatalf("Expected a value but got an error: %s", err)
-		}
-
-		if actual.Name != v.Expected.Name {
-			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
-		}
-
-		if actual.ResourceGroup != v.Expected.ResourceGroup {
-			t.Fatalf("Expected %q but got %q for Resource Group", v.Expected.ResourceGroup, actual.ResourceGroup)
-		}
-	}
-
-}
-
 func TestParseRelayNamespaceID(t *testing.T) {
 	testData := []struct {
 		Name     string
